(PC-13158)[API]refactor:moved sendinblue dataclasses no related to template
