+++ api/src/pcapi/core/mails/models/models.py
