+++ api/src/pcapi/core/mails/transactional/bookings/booking_cancellation_by_beneficiary.py
@@ -3,7 +3,7 @@ from typing import Union
 
 from pcapi.core import mails
 from pcapi.core.bookings.models import IndividualBooking
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.offers.utils import offer_app_link
 from pcapi.models.feature import FeatureToggle
