+++ api/src/pcapi/core/mails/transactional/bookings/booking_cancellation_by_institution.py
@@ -1,6 +1,6 @@
 from pcapi.core import mails
 from pcapi.core.educational.models import EducationalBooking
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
