+++ api/src/pcapi/core/mails/transactional/bookings/booking_confirmation_to_beneficiary.py
@@ -5,7 +5,7 @@ from pcapi.core.bookings.constants import BOOKINGS_AUTO_EXPIRY_DELAY
 from pcapi.core.bookings.constants import BOOKS_BOOKINGS_AUTO_EXPIRY_DELAY
 from pcapi.core.bookings.models import IndividualBooking
 from pcapi.core.categories import subcategories
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.models.feature import FeatureToggle
 from pcapi.utils.date import get_date_formatted_for_email
