+++ api/src/pcapi/core/mails/transactional/bookings/booking_postponed_by_pro_to_beneficiary.py
@@ -4,7 +4,7 @@ from babel.dates import format_date
 
 from pcapi.core import mails
 from pcapi.core.bookings.models import Booking
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.models.feature import FeatureToggle
 from pcapi.utils.mailing import format_booking_hours_for_email
