+++ api/src/pcapi/core/mails/transactional/pro/offer_validation_to_pro.py
@@ -2,7 +2,7 @@ from typing import Union
 
 from pcapi import settings
 from pcapi.core import mails
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import OfferValidationStatus
