+++ api/src/pcapi/core/mails/transactional/sendinblue_template_ids.py
@@ -1,40 +1,8 @@
-import dataclasses
 from enum import Enum
-from typing import Optional
 
-from pcapi import settings
-
-
-@dataclasses.dataclass
-class SenderInfo:
-    email: str
-    name: str
-
-
-class SendinblueTransactionalSender(Enum):
-    SUPPORT = SenderInfo(settings.SUPPORT_EMAIL_ADDRESS, "pass Culture")
-    SUPPORT_PRO = SenderInfo(settings.SUPPORT_PRO_EMAIL_ADDRESS, "pass Culture")
-    COMPLIANCE = SenderInfo(settings.COMPLIANCE_EMAIL_ADDRESS, "pass Culture")
-
-
-@dataclasses.dataclass
-class Template:
-    id_prod: int
-    id_not_prod: int
-    tags: list[str] = dataclasses.field(default_factory=list)
-    use_priority_queue: bool = False
-
-    # Tag your emails to find them more easily cf doc https://developers.sendinblue.com/reference/sendtransacemail
-    sender: SendinblueTransactionalSender = SendinblueTransactionalSender.SUPPORT
-
-    @property
-    def id(self) -> int:
-        return self.id_prod if settings.IS_PROD else self.id_not_prod
-
-
-@dataclasses.dataclass
-class TemplatePro(Template):
-    sender: SendinblueTransactionalSender = SendinblueTransactionalSender.SUPPORT_PRO
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalSender
+from pcapi.core.mails.models.sendinblue_models import Template
+from pcapi.core.mails.models.sendinblue_models import TemplatePro
 
 
 class TransactionalEmail(Enum):
@@ -133,23 +101,3 @@ class TransactionalEmail(Enum):
     REFUND_TO_PRO = TemplatePro(id_prod=0000, id_not_prod=0000, tags=[""])
     RESET_PASSWORD_TO_PRO = TemplatePro(id_prod=364, id_not_prod=47, tags=["pro_reinit_mdp"])
     WELCOME_TO_PRO = TemplatePro(id_prod=0000, id_not_prod=0000, tags=["pro-bienvenue-sur-le-pass"])
-
-
-@dataclasses.dataclass
-class SendinblueTransactionalEmailData:
-    template: Template
-    params: dict = dataclasses.field(default_factory=dict)
-
-
-@dataclasses.dataclass
-class SendinblueTransactionalAttachment:
-    content: str
-    name: str
-
-
-@dataclasses.dataclass
-class SendinblueTransactionalWithoutTemplateEmailData:
-    subject: str
-    html_content: str
-    sender: SendinblueTransactionalSender = SendinblueTransactionalSender.SUPPORT_PRO
-    attachment: Optional[SendinblueTransactionalAttachment] = None
