+++ api/src/pcapi/core/mails/transactional/users/email_address_change_confirmation.py
@@ -4,7 +4,7 @@ from typing import Union
 from dateutil.relativedelta import relativedelta
 
 from pcapi.core import mails
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.payments import conf as deposit_conf
 from pcapi.core.users import models as users_models
