+++ api/src/pcapi/core/mails/transactional/users/pre_subscription_dms_error.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from pcapi.core import mails
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
