+++ api/src/pcapi/core/mails/transactional/users/recredit_to_underage_beneficiary.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 
 from pcapi.core import mails
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.users import models as users_models
 from pcapi.core.users.api import get_domains_credit
