+++ api/src/pcapi/core/mails/transactional/users/reset_password.py
@@ -1,7 +1,7 @@
 from typing import Union
 
 from pcapi.core import mails
-from pcapi.core.mails.transactional.sendinblue_template_ids import SendinblueTransactionalEmailData
+from pcapi.core.mails.models.sendinblue_models import SendinblueTransactionalEmailData
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.users import api as users_api
 from pcapi.core.users.models import Token
